requirements.txt
asgiref==3.2.7
astroid==2.4.2
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2-binary==2.8.5
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.1.0
wrapt==1.12.1
